[refactor] 검색로직 JPQL 사용 및 리팩토링
