--- company/src/main/java/com/hubEleven/company/domain/repository/CompanySearchCondition.java
@@ -1,8 +1,0 @@
-package com.hubEleven.company.domain.repository;
-
-import com.hubEleven.company.domain.vo.CompanyStatus;
-import com.hubEleven.company.domain.vo.CompanyType;
-import java.util.UUID;
-
-public record CompanySearchCondition(
-		UUID hubId, String companyName, CompanyType type, CompanyStatus status) {}
